requirements.txt
# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0  # For parquet support

# Financial Data APIs
yfinance>=0.2.28
secedgar>=0.3.0
requests>=2.31.0

# NLP & ML
sentence-transformers>=2.2.0
vaderSentiment>=3.3.2
faiss-cpu>=1.7.4  # For vector search

# Utilities
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0

# AI/LLM
openai>=1.0.0
docetl>=0.1.0

# Supabase
supabase>=2.0.0